backend/pyproject.toml
[project]
name = "legal-rag-backend"
version = "0.2.0"
description = "Legal Intelligence Platform — Backend API"
requires-python = ">=3.11"
license = { text = "MIT" }

dependencies = [
    # --- Web framework ---
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "sse-starlette>=2.1.0",
    "pydantic>=2.9.0",
    "pydantic-settings>=2.6.0",

    # --- LLM (DeepSeek + OpenAI fallback) ---
    "openai>=1.55.0",

    # --- Embedding (Voyage AI) ---
    "voyageai>=0.3.0",

    # --- Vector DB (Qdrant) ---
    "qdrant-client>=1.12.0",

    # --- Reranker (local cross-encoder) ---
    "sentence-transformers>=3.3.0",
    "torch>=2.4.0",

    # --- Vietnamese NLP ---
    "underthesea>=6.8.0",

    # --- Document parsing ---
    "llama-parse>=0.5.0",
    "pymupdf>=1.24.0",
    "python-docx>=1.1.0",
    "beautifulsoup4>=4.12.0",
    "lxml>=5.3.0",

    # --- PostgreSQL ---
    "sqlalchemy[asyncio]>=2.0.0",
    "asyncpg>=0.30.0",
    "alembic>=1.14.0",

    # --- Auth ---
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.0",

    # --- Cache & sessions (Redis) ---
    "redis[hiredis]>=5.2.0",

    # --- HTTP client ---
    "httpx>=0.28.0",

    # --- Utilities ---
    "python-multipart>=0.0.17",
    "pyyaml>=6.0",
    "python-dotenv>=1.0.0",
    "numpy>=1.26.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
    "httpx>=0.28.0",
    "ruff>=0.8.0",
    "mypy>=1.13.0",
    "pre-commit>=4.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.mypy]
python_version = "3.11"
strict = true
